requirements.txt
beautifulsoup4==4.12.3
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchainhub==0.1.21
langsmith==0.2.7
lxml==5.3.0
playwright==1.49.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
pydantic-settings==2.7.1